+++ projects/ui/src/Components/ApiDetails/ApiSchemaDisplay.tsx
@@ -2,8 +2,8 @@ import { Button } from "@mantine/core";
 import { useMemo, useState } from "react";
 import { di } from "react-magnetic-di";
 import { useParams } from "react-router-dom";
-import YAML from "yaml";
 import { useGetApiDetails, useListApis } from "../../Apis/hooks";
+import { createBackstageYaml } from "../../Utility/backstage_demo_utility";
 import { downloadFile } from "../../Utility/utility";
 import { Loading } from "../Common/Loading";
 import { RedocDisplay } from "./RedocDisplay";
@@ -11,19 +11,6 @@ import { SwaggerDisplay } from "./SwaggerDisplay";
 
 // This is a flag to enable the backstage yaml download button.
 const BACKSTAGE_YAML_DEMO_ENABLED = false;
-
-function keyValue(theKey: string, ...theValues: (string | undefined)[]) {
-  let truthyValue = "";
-  for (let i = 0; i < theValues.length; i++) {
-    const cur = theValues[i];
-    if (cur !== undefined) {
-      truthyValue = cur;
-      break;
-    }
-  }
-  if (!truthyValue) return "";
-  else return theKey + ": " + truthyValue;
-}
 
 /**
  * MAIN COMPONENT
@@ -34,31 +21,14 @@ export function ApiSchemaDisplay() {
   const { isLoading: isLoadingApiDetails, data: apiSchema } =
     useGetApiDetails(apiId);
   const { isLoading: isLoadingApisList, data: apisList } = useListApis();
+  const isLoading = isLoadingApiDetails || isLoadingApisList;
 
   const backstageYaml = useMemo(() => {
     if (!BACKSTAGE_YAML_DEMO_ENABLED) return undefined;
     const apiSummary = apisList?.find((a) => a.apiId === apiId);
     if (!apiSummary || !apiSchema) return undefined;
-    const meta = apiSummary.customMetadata;
-    const apiSchemaYamlStr = YAML.stringify(apiSchema);
-    return `apiVersion: backstage.io/v1alpha1
-kind: API
-metadata:
-  name: ${apiSummary.title.replaceAll(" ", "-")}
-  description: ${apiSummary.description}
-spec:
-  type: openapi
-  ${keyValue("owner", meta?.owner, apiSummary.contact)}
-  ${keyValue("lifecycle", meta?.lifecycle, "production")}
-  definition: |-
-    ${
-      // Add in the indents.
-      apiSchemaYamlStr.replaceAll("\n", "\n    ")
-    }
-`;
+    return createBackstageYaml(apiSummary, apiSchema);
   }, [apisList, apiSchema]);
-
-  const isLoading = isLoadingApiDetails || isLoadingApisList;
 
   const [isSwagger, setIsSwagger] = useState(false);
